// frontend/src/types.ts
export interface ClientData {
    ageNow: number;
    ageRetirement: number;
    ageDeath: number;
    c0: number;
    p: number;
    r: number;
    g: number;
    i: number;
  }
  
  export interface TableRowData {
    age: number;
    capitalYearStart: number;
    interest: number;
    saved: number;
    capitalYearEnd: number;
    savedPpMonthly: number;
  }